#93 Refactor: 서비스 계층 인터페이스/구현 분리 및 impl 패키지 도입
